Tailwind version

+++ app/Http/Controllers/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Category;
 use \Cviebrock\EloquentSluggable\Services\SlugService;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller
 {
@@ -15,11 +16,18 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        if(!Gate::allows('admin')) {
+            abort(403);
+        };
+
         $categories = Category::all();
+
+        $count = Category::all()->count();
 
         return view('operator.categories.index', [
             'title' => 'Categories',
-            'categories' => $categories
+            'categories' => $categories,
+            'count' => $count
         ]);
     }
 
@@ -30,6 +38,10 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        if(!Gate::allows('admin')) {
+            abort(403);
+        };
+
         return view('operator.categories.create', [
             'title' => 'Categories'
         ]);
@@ -43,6 +55,10 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
+        if(!Gate::allows('admin')) {
+            abort(403);
+        };
+
         // Validation Input
         $request->validate([
             'name' => 'required'
@@ -55,7 +71,7 @@ class CategoryController extends Controller
             'description' => $request->description,
         ]);
 
-        return redirect('categories')->with('alert-success', 'Create New Category Success!');
+        return redirect('categories')->with('status-success', 'Create New Category Success!');
     }
 
     /**
@@ -77,6 +93,10 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        if(!Gate::allows('admin')) {
+            abort(403);
+        };
+
         return view('operator.categories.edit', [
             'title' => 'Categories',
             'category' => $category
@@ -92,6 +112,10 @@ class CategoryController extends Controller
      */
     public function update(Request $request, Category $category)
     {
+        if(!Gate::allows('admin')) {
+            abort(403);
+        };
+
         // Validate input
         $request->validate([
             'name' => 'required'
@@ -111,7 +135,7 @@ class CategoryController extends Controller
         // Update data
         Category::where('id', $category->id)->update($data);
 
-        return redirect('categories')->with('alert-success', 'Update Category Success!');
+        return redirect('categories')->with('status-success', 'Update Category Success!');
     }
 
     /**
@@ -122,14 +146,18 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
+        if(!Gate::allows('admin')) {
+            abort(403);
+        };
+
         // Check if the category is connected with several posts 
         try {
             Category::destroy($id);
         } catch (\Throwable $th) {
-            return redirect('categories')->with('alert-danger', 'Category Cannot Be Deleted! : This Category is Connected with Several Posts');
+            return redirect('categories')->with('status-danger', 'Category Cannot Be Deleted! : This Category is Connected with Several Posts');
         }
 
-        return redirect('categories')->with('alert-success', ' Category Has Been Deleted!');
+        return redirect('categories')->with('status-success', ' Category Has Been Deleted!');
     }
 
     /**
